rename file and fix reviewed comments
