+++ src/components/status-pipeline/LabelStatus.tsx
@@ -1,17 +1,18 @@
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
+
 import Typography from "@/components/core/typography";
 import rem from "@/utils/rem";
 import { customStatusStyle, statusIconStyle } from "./styles";
 
-import type { StatusPipelineProps } from "./types";
+import type { LabelStatusProps } from "./types";
 
-export default function StatusPipeline({
+export default function LabelStatus({
   status = "ready",
   text,
-}: StatusPipelineProps) {
+}: LabelStatusProps) {
   return (
-    <Stack direction={"row"} gap={rem(4)} alignItems="center">
+    <Stack direction="row" gap={rem(4)} alignItems="center">
       <Box
         sx={{
           backgroundColor: customStatusStyle[status].color,
