--- src/components/status-pipeline/StatusPipeline.stories.tsx
@@ -1,17 +1,0 @@
-import StatusPipeline from "./StatusPipeline";
-import type { Meta, StoryObj } from "@storybook/react";
-
-const meta: Meta<typeof StatusPipeline> = {
-  component: StatusPipeline,
-  title: "Components/StatusPipeline",
-};
-
-export default meta;
-
-type Story = StoryObj<typeof StatusPipeline>;
-
-export const Default: Story = {
-  args: {
-    status: "ready",
-  },
-};
