+++ src/components/status-pipeline/StatusPipeline.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from "@testing-library/react";
-import StatusPipeline from "./StatusPipeline";
+import StatusPipeline from "./LabelStatus";
 import { customStatusStyle } from "./styles";
 
 describe("StatusPipeline Component", () => {
-  it("should render with default status 'ready' when no props are provided", () => {
+  it("renders with default status 'ready' when no props are provided", () => {
     render(<StatusPipeline />);
 
     // Check that the text defaults to 'ready'
@@ -16,14 +16,14 @@ describe("StatusPipeline Component", () => {
     );
   });
 
-  it("should render the correct text when provided", () => {
+  it("renders the correct text when provided", () => {
     render(<StatusPipeline text="Custom Text" status="running" />);
 
     // Check that the custom text is rendered
     expect(screen.getByText(/custom text/i)).toBeInTheDocument();
   });
 
-  it("should render the correct status styles based on the 'status' prop", () => {
+  it("renders the correct status styles based on the 'status' prop", () => {
     render(<StatusPipeline status="completed" />);
 
     // Check that the text defaults to the 'completed' status
@@ -36,7 +36,7 @@ describe("StatusPipeline Component", () => {
     );
   });
 
-  it("should apply uppercase style to the Typography", () => {
+  it("applies uppercase style to the Typography", () => {
     render(<StatusPipeline status="failed" />);
 
     const typographyElement = screen.getByText(/failed/i);
